requirements.txt
drf_yasg==1.17.1
Django==3.0.6
djangorestframework_jwt==1.11.0
djangorestframework==3.11.0
